chore(licenses): remove dead dependencies
JSON.net otherwise known as Newsonsoft.Json is not going to be used anymore as .NET as a good enough replacement for it.

ImageListView is related to WPF and doesn't work without it, so... :3

While they may be referenced by ShareX.csproj, that project can't build and isn't distributed anywhere.