fix(NMH): provide an actual description for web browser extensions
This commit also aligns the Chrome application id to the name, preventing bugs. There was another part of this work I was working on that'd fix the Chrome extension completely, but, it's on my gaming laptop and I may have left the charger at home... I'm currently on holiday vacation.