fix(Core): enable reflection until migration is complete
This commit also includes a new building system (NUKE) that will be used in the future to shift variables for artifacts so that they're flexible for Linux distributions

This commit also fixes a bug in the FindSnapX function where it was looking for binaries with no file extension which makes no sense. Now it's looking for binaries with .exe by default.