refactor(ScreenCapture): finish refactor
Of course ScreenCaptureLib doesn't work anymore but I am looking to add back FFMpeg and the image effects.
I sunk everything related to shapes as it's a bastion of bad code principles with too much of a reliance on WPF,System.Drawing, and the separation of concerns is very poor. It cannot be reformed.