ci(build): build on Ubuntu 24.04 ARM
This will provide AARCH64 architecture builds for SnapX on Linux
Thanks to github/roadmap#970
Hopefully actions/runner-images#10820 will be added and then Windows on ARM support will follow. Fingers crossed, #33

Additionally, the artifact `SnapX-Vendored-Source` has been renamed to `SnapX-Vendored-Source-${{ runner.arch }}` to accomodate for architectural binaries. What that means is that I shouldn't have left it not suffixed with the runner's architecture in the name in the first place.