feat(Core): add broken TesseractOCR support
Listen, I tested this with NativeAOT off. It works, but, if NativeAOT is on, dynamic code generation is illegal. I am working with TesseractOCR project lead to fix this. For now, the code is there. It works when NativeAOT is off. Just for reference all of our builds use NativeAOT. Fuck