build: make .NET version authoritative
Previously, if I didn't, people would have problems with the rollout of .NET 9 because it hadn't released for their distro. Now, that's not a concern because of NUKE automatically downloading the specified version when .NET is not installed or outdated. For Linux distros that patch .NET, it might be. But, that's not my problem :)